Move formatEpochTime to a shared time util
formatEpochTime is a generic Matter-epoch timestamp formatter, but it
lived in util/commodity-tariff.ts and was imported by the Thermostat
panel too, coupling an unrelated feature to the tariff module. Move it
to util/time.ts and import from there instead.
